Creates a random maze that won't cause any loops.
(This code is an adaption from a stackoverflow user "DoubleMx2". 

It derives heavily from it with difference only being in variable name changes.
To stay coherent to our already established naming scheme we've excluded
the toString and main methods that is present in the original code.
  
credit: http://stackoverflow.com/questions/18396364/maze-generation-arrayindexoutofboundsexception